requirements.txt
newsapi_python==0.2.7
google-cloud-storage==2.14.0